implements logic to log the user in on install
